Create separate DataSegment types for webcil and nativeaot data segments.
