+++ src/coreclr/tools/aot/ILCompiler.Compiler/ILCompiler.Compiler.csproj
@@ -380,7 +380,9 @@
     <Compile Include="..\..\Common\Compiler\ObjectWriter\WasmRelocatableObjectWriter.cs" Link="Compiler\ObjectWriter\WasmRelocatableObjectWriter.cs" />
     <Compile Include="..\..\Common\Compiler\ObjectWriter\WasmObjectWriter.cs" Link="Compiler\ObjectWriter\WasmObjectWriter.cs" />
     <Compile Include="..\..\Common\Compiler\ObjectWriter\Wasm\WasmDataSection.cs" Link="Compiler\ObjectWriter\Wasm\WasmDataSection.cs" />
-    <Compile Include="..\..\Common\Compiler\ObjectWriter\Wasm\WasmDataSegment.cs" Link="Compiler\ObjectWriter\Wasm\WasmDataSegment.cs" />
+    <Compile Include="..\..\Common\Compiler\ObjectWriter\Wasm\WasmDataSegmentEmitter.cs" Link="Compiler\ObjectWriter\Wasm\WasmDataSegmentEmitter.cs" />
+    <Compile Include="..\..\Common\Compiler\ObjectWriter\Wasm\WasmByteArrayDataSegment.cs" Link="Compiler\ObjectWriter\Wasm\WasmByteArrayDataSegment.cs" />
+    <Compile Include="..\..\Common\Compiler\ObjectWriter\Wasm\IWasmDataSegment.cs" Link="Compiler\ObjectWriter\Wasm\IWasmDataSegment.cs" />
     <Compile Include="..\..\Common\Compiler\ObjectWriter\Wasm\WasmSection.cs" Link="Compiler\ObjectWriter\Wasm\WasmSection.cs" />
     <Compile Include="..\..\Common\Compiler\ObjectWriter\Wasm\WasmSections.cs" Link="Compiler\ObjectWriter\Wasm\WasmSections.cs" />
     <Compile Include="..\..\Common\Compiler\ObjectWriter\WasmNative.cs" Link="Compiler\ObjectWriter\WasmNative.cs" />
